requirements.txt
fastapi==0.89.1
uvicorn[standard]
regex==2022.6.2
urllib3==1.26.14
urlextract==1.5.0
validators==0.20.0
pandas==1.4.3
requests==2.28.2
tldextract==3.4.0
bs4
scikit-learn==1.2.1
numpy==1.24.3
tld==0.13
fastapi-cache2==0.2.1
redis==4.6.0
